03-assignments/01-assignment-chatbot/requirements.txt
groq
pydantic
python-dotenv
langchain
langchain-community 
langchain-core
langchain-groq
jsonschema
langchain_core.output_parsers JsonOutputParser